# docs/swagger.yaml
basePath: /
definitions:
  models.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
...
  server.loginReq:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  server.loginResDocs:
    properties:
      accessToken:
        type: string
      user:
        $ref: '#/definitions/server.userDocs'
    type: object
...
  server.userDocs:
    properties:
      PhoneNumber:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      image:
        type: string
      lastName:
        type: string
      role:
        $ref: '#/definitions/models.Role'
    type: object
...
  utils.APIError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: ali.jabar@refinesoftware.net
    name: Ali Jabar
  description: Afrad is a RESTful API for a clothing e-commerce platform.
  license:
    name: MIT License
    url: https://img.shields.io/badge/License-MIT-yellow.svg
  title: Afrad API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user using email and password. Returns an access token
        and user data.
      parameters:
      - description: Login request
        in: body
        name: loginReq
        required: true
        schema:
          $ref: '#/definitions/server.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login with access token and user info
          schema:
            $ref: '#/definitions/server.loginResDocs'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Invalid credentials or unverified account
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Email/Password Login
      tags:
      - Auth
...
  /oauth/google/callback:
    get:
      consumes:
      - application/json
      description: Handles the Google OAuth callback, authenticates the user, and
        returns a JWT access token.
      parameters:
      - description: OAuth authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state (if used)
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful login with JWT token and user data
          schema:
            $ref: '#/definitions/server.loginResDocs'
        "400":
          description: Bad request or invalid input
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized - Invalid OAuth token
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Google OAuth Callback
      tags:
      - OAuth
  /oauth/google/login:
    get:
      consumes:
      - application/json
      description: Redirects the user to Google's OAuth 2.0 login screen.
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to Google
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Start Google OAuth Login
      tags:
      - OAuth
  /user/logout:
    post:
      description: Logs out the currently authenticated user by revoking the session
        and clearing the refresh token cookie.
      produces:
      - application/json
      responses:
        "204":
          description: Successfully logged out
        "400":
          description: Missing refresh token or invalid request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized or invalid session
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - User
  /user/logout/all:
    post:
      description: Revokes all active sessions for the authenticated user across all
        devices.
      produces:
      - application/json
      responses:
        "204":
          description: Successfully logged out from all sessions
        "401":
          description: Unauthorized or invalid token
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Logout from All Sessions
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your_jwt_token}" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"